handle submit on page
